refactor: 발행한 로또 번호 출력 기능 Lotto 클래스 사용
